:recycle: refactor: Cria um componente de Input para deixar o código mais limpo e legível
